LayoutTests/inspector-obsolete/http-tests/network/network-xhr-async-double.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script src="../network-test.js"></script>
<script>
function test()
{
    // Since this test could be run together with other inspector backend cache
    // tests, we need to reset size limits to default ones.
    InspectorTest.resetInspectorResourcesData(step1);

    function step1()
    {
        InspectorTest.makeSimpleXHRWithPayload("POST", "resources/echo-payload.php", true, 'request1', step2);
    }

    function step2()
    {
        InspectorTest.makeSimpleXHRWithPayload("POST", "resources/echo-payload.php", true, 'request2', step3);
    }

    function step3()
    {
        var request1 = WebInspector.panel("network").requests[WebInspector.panel("network").requests.length - 2];
        var request2 = WebInspector.panel("network").requests[WebInspector.panel("network").requests.length - 1];
        request1.requestContent(step4);
        request2.requestContent(step4);
    }

    var toLoad = 2;
    function step4()
    {
        if (--toLoad)
            return;
        var request1 = WebInspector.panel("network").requests[WebInspector.panel("network").requests.length - 2];
        var request2 = WebInspector.panel("network").requests[WebInspector.panel("network").requests.length - 1];
        InspectorTest.addResult("resource1.content: " + request1.content);
        InspectorTest.addResult("resource2.content: " + request2.content);
        InspectorTest.assertTrue(request1.content === 'request1' && request2.content === 'request2');
        InspectorTest.completeTest();
    }
}
</script>
</head>
<body onload="runTest()">
Tests responses in network tab for two XHRs sent without any delay between them.
<a href="https://bugs.webkit.org/show_bug.cgi?id=91630">Bug 91630</a>
</body>
</html>
