LayoutTests/inspector-obsolete/http-tests/network/network-xhr-async-double-expected.txt

CONSOLE MESSAGE: line 6: XHR loaded: 1
CONSOLE MESSAGE: line 6: XHR loaded: 2
Tests responses in network tab for two XHRs sent without any delay between them. Bug 91630
resource1.content: request1
resource2.content: request2
